Rename lnx::periph to lnx
