+++ firmware/projects/FrontController/platforms/linux/bindings.cc
@@ -5,13 +5,13 @@
 #include <chrono>
 #include <iostream>
 
-#include "mcal/linux/periph/analog_input.hpp"
-#include "mcal/linux/periph/can.hpp"
-#include "mcal/linux/periph/digital_input.hpp"
-#include "mcal/linux/periph/digital_output.hpp"
+#include "mcal/linux/analog_input.hpp"
+#include "mcal/linux/can.hpp"
+#include "mcal/linux/digital_input.hpp"
+#include "mcal/linux/digital_output.hpp"
 
 namespace mcal {
-using namespace lnx::periph;
+using namespace lnx;
 
 // =========== CAN =========================================
 CanBase veh_can_base{"vcan0", false};
